Add test_contexts
